+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/logic/versions/deneb/helpers/MiscHelpersDeneb.java
@@ -99,20 +99,25 @@ public class MiscHelpersDeneb extends MiscHelpersBellatrix {
               blobSidecar.getBlockRoot(),
               beaconBlockRoot);
         });
-    checkArgument(
-        kzgCommitments.size() == blobSidecars.size(),
-        "Number of KZG commitments (%s) does not match number of blobs (%s)",
-        kzgCommitments.size(),
-        blobSidecars.size());
     final List<Bytes> blobs =
         blobSidecars.stream()
             .map(BlobSidecar::getBlob)
             .map(Blob::getBytes)
             .collect(Collectors.toList());
-    final List<KZGProof> kzgProofs =
+    final List<KZGProof> proofs =
         blobSidecars.stream().map(BlobSidecar::getKZGProof).collect(Collectors.toList());
+    checkArgument(
+        kzgCommitments.size() == blobs.size(),
+        "Number of KZG commitments (%s) does not match number of blobs (%s)",
+        kzgCommitments.size(),
+        blobSidecars.size());
+    checkArgument(
+        blobs.size() == proofs.size(),
+        "Number of blobs (%s) does not match number of proofs (%s)",
+        blobs.size(),
+        proofs.size());
     checkState(
-        kzg.verifyBlobKzgProofBatch(blobs, kzgCommitments, kzgProofs),
+        kzg.verifyBlobKzgProofBatch(blobs, kzgCommitments, proofs),
         "The blobs and KZG proofs do not correspond to the KZG commitments for slot %s and block root %s",
         slot,
         beaconBlockRoot);
